trisoup/m61577: non-cubic nodes
Add possibility to use non cubic nodes on boundaries of the slices in
order to avoid reconstruction artefacts and discontinuities between
neighbouring slices.
Node cubic nodes are determined based on a bounding box provided for the
trisoup volume within the slice.

In can be enabled separately for lowest coordinates in the slice and
highest ones.

Encoder decision is also added to determine the bounding box parameters.