trisoup: m63661 - inter skip mode
In non-moving parts of a dynamic point cloud, it has been observed that
the geometry inter prediction based on (zero) motion compensation may
not be optimal due to lack of invariance of the compensation process.

A new inter coding mode based on colocated vertices and nodes has been
introduced to mimic a kind of skip mode for point cloud coding.