raht: fix - use safer approximation of tree depth
Method used for estimating raht tree depth in m66238 does not
look safe: the distance between the first an last point in morton
order is not representative of the actual boundaries of the point
cloud. Also the function roundLog2 does not look correct.

This is fixed by only considering the last point in morton order.