geom/attr: refactor - payload handling structure for single pass decoding
This commit refactors the decoding process to allow expected feature of
being able to perform geometry and attributes decoding together in a
single pass (with more local processing of the data).

Payloads are not directly decoded, they are stored together to behave
like a decoder that would decode them in parallel or in a single pass.

Decoding effectively starts when the end of a frame is detected, and the
geometry as well as all the attributes are decoded "simultaneously".