trisoup: m67015 - support for non-powers of two node sizes
This commit provides finer granularity for quality control of the
TriSoup model by introducing the support for TriSoup node sizes which
are not powers of two.

The interface with octree structures in intra and inter coding is based
on dicing and undicing the point clouds. It allows to keep powers of two
nodes within the octrees to not impact the complexity and the execution
speed.

integrated from commit 02e9f4fe1e8b406d69438ee1c59a1d01ca76766f.