utils: dynamic reallocation for encoding buffers
This commits allows dynamic reallocation of arithmetic encoding
buffers in case the preallocated size is too small.

It avoids encoder crashes if the initial coding buffer size was too
small. However a limit of 512 MB is put for dynamic reallocation to
avoid exhausting the whole system memory in case of an encoder bug.