raht: m67627 - fix prediction mode signaling
Prediction mode is signaled for every layer in its specification, but
the signaling was not performed for the first one.

This commit modifies the software to match with the spec when signalling
the prediction mode.

integrated from commit e4195c9c7d6d25a1018d9e75decd7d52ae1e0625