enc: fix issue #5 - do not use std::log2 for RDO
This commit removes dependencies on std::log2() for the encoder when
performing Rate Distortion Optimizations (RDO). Instead, a fixed point
approximation of log2 pcc::fpLog2 is introduced so that the results will
not depend on the platform and its implementation dependent math library.

This commit thus changes computations for RAHT RDO, as well as for motion
search, and should solve issue #5.