raht: fix issue #9 - RDO is not good for root node
When a root node is coded with raht, the rdo is not taking into account
the rate/distorsion for the DC coefficients. Furthermore on the root
node the lambda is always zero, totally ignoring the rate cost.

This is fixed by including DC coefficients for the root node,
and updating R/D statistics before computing lambda in order to not
have it null.