raht: fix issue #6 - decoder crash on determining neighbors mode
The decoder may crash when decoding a bitstream generated by an encoder
configured with --rahtUpperModeLevel=0 or --rahtUpperModeLevel=1.

The crash is caused by the decoder trying to determine neighbors/uncles'
mode for the root node based on unitialized neighbor indexes.

The neighbor indexes are only available when intraPred or intraLayerPred
is enabled. The bug is then simply fixed by imposing the neighborsMode to
be Inter when intraPred is not enabled (or Null if interPred is not
enabled).