raht: fix issue #7 - encoder/decoder mismatch on coefficients values
In then encoder, when rahtUpperModeLevel is set to 0 (or 1) with inter
prediction enabled and with layer mode coding enabled, the first layer is
not infered and there is a copy needed to retrieve the coefficients from
the inter layer coefficients buffer.

On firt layer the DC coefficient was not comptabilized in the number of
coefficients to copy. The buffer size was wrong and one coefficient was
missing in the copy, leading to a mismatch with the decoder.

This is fixed by setting the number of coefficients to the correct value
to get the right buffer size and number of copied coefficients.