raht: fix issue #8 - encoder/decoder mismatch with average prediction
On the encoder side, when using inter prediction with average prediction
enabled and layer mode prediction enabled, the encoder could modify the
inter predictor if the averageLayerPrediction was enabled on a layer but
not the intraPrediction (this may append on the second layer, if
rahtUpperModeLevel is set to 0, 1 or 2).

On decoder side, the predictor is not averaged since it is not allowed
to perform intra prediction on that layer.

The resulting missmatch is solved by enforcing the encoder to only modify
the inter predictor if the intra prediction is also enabled on the layer.