raht: major refactoring
This commits continues the major refactoring of the RAHT code.
It continues the work started in
commit 9020c316b2231ac3012cdf84d392365d4816ed1f

Most of the changes that were made in the decoder have been applied to
the encoder in a similar way, when possible, to fit the RDO nature of the
encoder.

Some of the changes are reported bellow:
 - encoder uses similar bottom-up process as already implemented in
   decoder
 - inter predictor computation is moved into bottom-up process (as
   geometry of the inter predictor is now always the same as the
   reconstructed geometry of the current frame)
 - translateLayer and interTree are removed, since now useless
 - bottom-up process is now templatized

The encoder's RDO process has been optimized by removing most of the
Lagrangian costs computations made in double, they are now performed
using integers. The fixed point fpLog2 approximation of log2
-- which was only applied to probabilities, i.e to the (0; 1] interval,
   for entropy/bits cost computation --
has been replaced by a linear approximation from the tabulated interval.

Other RDO/encoder optimizations have been applied, like:
 - storing of pred mode in child nodes,
 - getting the prediction mode,
 - ...

Various additional code compactions, optimizations and cleanups have
been applied to either the encoder, the decoder or both.

Additional Notes:

In the current implementation the handling of the duplicated points is
broken in the decoder, because of the simplifications introduced for
Haar nodes (weights are not needed in these nodes). This could be fixed
for instance by storing the duplicated counts in a dedicated array, but
one question to answer first is: "Is it still relevant to handle
duplicated points within the GeS-TM ?"
If yes we could fix the issue; if no the codec could be further
simplified.

There are a few normative changes that were needed:
 - inter predictor was using 3 integer divisions which have been replaced
   by an integer division approximated from the approximated integer
   inverse square root: 1/x = 1/(sqrt(x)*sqrt(x)) ~ irsqrt(x)*irsqrt(x)
 - inter predictor for Haar was a RAHT predictor and has been replaced by
   a Haar predictor (fix).

The call to scanBlock has been inlined and the traversal order
of the coefficients kRahtScanOrder has been removed (which could possibly
have normative impacts)

Co-authored-by: Jonathan Taquet <jtaquet@ofinno.com>